cleanup and simplify generation codeclear
